configure: Add correct CFLAGS and LDFLAGS for libssl11
This is continuation of previous commit on OpenSSL 1.1
The new 1.1 version is needed for TLSv1.3 support.
By default centos7 uses openssl-devel version 1.0.2k.
The newer version 1.1.1k is available with EPEL repo.

From the naming perspective for pkg-config:
- Only Centos7 via EPEL calls it `libssl11`
- All the rest call it `libssl`

Even on Fedora with OpenSSL 3 is is called `libssl`.

See:
$ pkg-config --modversion libssl
1.0.2k
$ pkg-config --modversion libssl11
1.1.1k
$ pkg-config --modversion libssl
3.0.2

So it's safe to check for `libssl11` here

Change-Id: I60b161b166055b6aa4dc3b1f674a29cb7b3bb7c8
Signed-off-by: Boris Glimcher <Boris.Glimcher@emc.com>